+++ src/interfaces/resources/order-details.interface.ts
@@ -1,0 +1,16 @@
+import { OrderItemInterface } from "./order-item.interface";
+
+export interface OrderDetailsInterface {
+	readonly items: Array<OrderItemInterface>,
+	readonly order_contains_tickets_for_other_organizations: boolean
+}
+
+export const createOrderDetails = (base: any = {}): OrderDetailsInterface => {
+	return {
+		...{
+			items: [],
+			order_contains_tickets_for_other_organizations: false
+		},
+		...base
+	} as OrderDetailsInterface;
+};