+++ src/interfaces/resources/refund-response.interface.ts
@@ -1,0 +1,15 @@
+export interface RefundResponseInterface {
+	readonly amount_refunded: number,
+	readonly refund_breakdown: {[paymentMethod: string]: number}
+}
+
+
+export const createRefundResponse = (base: any = {}): RefundResponseInterface => {
+	return {
+		...{
+			amount_refunded: 0,
+			refund_breakdown: {}
+		},
+		...base
+	} as RefundResponseInterface;
+};